Move modules from utils.files to api._modules
The utils should not require the api, the api._modules module is there
just for the reason to create simple modules that are not required
anywhere else.
